2025.05.28
댓글, 대댓글 삭제 API 구현
User Entity 생성 완료
User 모듈 구조 작성 완료

+++ src/main/java/com/weatherfit/comment_service/comment/repository/CommentRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
 public interface CommentRepository
-        extends ReactiveCrudRepository<Comment, Integer> {
+        extends ReactiveCrudRepository<Comment, Long> {
 
     // boardId 로 Comment 리스트 조회
-    Flux<Comment> findByBoardId(Integer boardId);
+    Flux<Comment> findByBoardId(Long boardId);
 
 }
